YourCarbonScore_Quark/requirements.txt
asgiref==3.6.0
Django==3.2.17
django-social-share==2.3.0
joblib==1.2.0
numpy==1.24.2
python-dotenv==0.21.1
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.0
sqlparse==0.4.3
threadpoolctl==3.1.0
